fix: avoid mounting /usr/src for underlying Linux distros without it
Ubuntu has /usr/src with the kernel headers but on Flatcar Container
Linux, this directory does not exist and the kernel headers are located
in /lib/modules/`uname -r` directly.

Attempting to mount /usr/src from the host causes Docker to attempt to
create that directory but this does not work when the host /usr is
read-only.

This patch mounts the host /usr into the tracerunner /usr-host instead
because /usr always exists. Then, a symlink redirects /usr/src to
/usr-host/src when needed.

Signed-off-by: Alban Crequy <alban@kinvolk.io>