feat: Support generic tracers, bcc, target namespace
This adds support for generic tracers, fixing #125

The new "--tracer" flag allows specifying arbitrary generic tracers.

The bcc suite of tools is added to take advantage of this, as well as a "fake"
tracer for integration testing this functionality, fixing #112

The primary original author of this functionality was Zeeshan Qureshi.

In order to pull this change in, it was necessary to also pull in functionality
that refactored how trace job targets are specified. This introduces a new flag
which allows specifying the target namespace separately from the namespace
where the trace is created, which fixes #147. This functionality was co-authored
with Aaron Olson.

Co-authored-by: Aaron Olson <934893+honkfestival@users.noreply.github.com>
Co-authored-by: Zeeshan Qureshi <zee@zqureshi.in>