feat: Support selecting target process to trace with "process selector"
This adds the ability to target a specific process to trace with the new flag
`--process-selector`. The prior behavior was that only the root PID of a pod
would be traced, which tends to not be the desired target process in most
cases.

Especially with the addition of generic tracers which may only work when a
target pid is provided, the ability to have more fine-grained control over what
process within a pod is targeted is desirable.

The selected process is substituted for `$container_pid` to preserve
backward-compatibility.

The primary author of this functionality was Aaron Olson, but it built on
earlier work from Zeeshan Qureshi who implemented the preliminary selector
work.

Co-authored-by: Aaron Olson <934893+honkfestival@users.noreply.github.com>
Co-authored-by: Zeeshan Qureshi <zee@zqureshi.in>