feat: Support downloading trace output, uploading to GCS
This adds support for specifying the `--output` flag, which can either be:

- stdout (the existing behavior)
- A local path, resulting in the output being downloaded to a local location
- A URI for a GCS path, to upload to google blob storage

If the target is GCS, a corresponding application secret must be specified,
indicating a kubernetes secret that contains a service account key with
access to upload to GCS.

In all cases, the output will be tee'd to stdout as well as any desired
additional location.

The primary author of the upload functionality was Zeeshan Qureshi, this commit
is made on his behalf.

Co-authored-by: Zeeshan Qureshi <zee@zqureshi.in>
Co-authored-by: Aaron Olson <934893+honkfestival@users.noreply.github.com>