feat: Add deploy/deployment as target resource option
This adds support to target by deployment, rather than having to manually
specify a target pod.

This is a convenience feature, which will return the first valid pod within
a deployment, then use the existing pod-processing logic from there. A pod
is currently considered to be valid if it is on a node which is allocatable
(so we can actually schedule a trace job there).