fix: Provide more meaningful default version info
This change allows the version info to be inferred at build time to the best
extent possible.

If buildTime is omitted, it will default to the current time.
If the gitCommit is omitted, it will default to "unspecified", indicating that
this was not passed by an LD flag. The docs reflect how to properly set this
flag.